Use getSingleAssetSrc across codebase
