+++ src/features/vault/components/PoolSummary/PoolTitle/PoolTitle.js
@@ -5,27 +5,11 @@ import AvatarGroup from '@material-ui/lab/AvatarGroup';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
+import { getSingleAssetSrc } from '../../../../helpers/getSingleAssetSrc';
 
 import styles from './styles';
 
 const useStyles = makeStyles(styles);
-
-const singleAssetRequire = require.context('images/single-assets', false, /\.(svg|webp|png)$/);
-const singleAssets = Object.fromEntries(
-  singleAssetRequire.keys().map(path => [path.substring(2, path.lastIndexOf('.')), path])
-);
-const singleAssetCache = {};
-const singleAsset = symbol => {
-  if (symbol in singleAssetCache) {
-    return singleAssetCache[symbol];
-  }
-
-  if (symbol in singleAssets) {
-    return (singleAssetCache[symbol] = singleAssetRequire(singleAssets[symbol]));
-  }
-
-  throw new Error(`Image required for '${symbol}' token in 'images/single-assets/'`);
-};
 
 const PoolTitle = ({
   name,
@@ -59,13 +43,13 @@ const PoolTitle = ({
           alt={assets[0]}
           variant="square"
           imgProps={{ style: { objectFit: 'contain' } }}
-          src={singleAsset(assets[0])}
+          src={getSingleAssetSrc(assets[0])}
         />
         <Avatar
           alt={assets[1]}
           variant="square"
           imgProps={{ style: { objectFit: 'contain' } }}
-          src={singleAsset(assets[1])}
+          src={getSingleAssetSrc(assets[1])}
         />
       </AvatarGroup>
     );
